Moved user behavior to another protocol
