+++ Amplify/Categories/Auth/AuthCategoryBehavior.swift
@@ -11,7 +11,7 @@ import UIKit
 public typealias AuthUIPresentationAnchor = UIWindow
 
 /// Behavior of the Auth category that clients will use
-public protocol AuthCategoryBehavior {
+public protocol AuthCategoryBehavior: AuthCategoryUserBehavior {
 
     /// SignUp a user with the authentication provider.
     ///
@@ -125,34 +125,4 @@ public protocol AuthCategoryBehavior {
                                listener: AuthConfirmForgotPasswordOperation.EventListener?) ->
     AuthConfirmForgotPasswordOperation
 
-    // MARK: - User Management
-
-    /// Returns the currently logged in user.
-    ///
-    func getCurrentUser() -> AuthUser?
-
-    func fetchAttributes(options: AuthFetchUserAttributeOperation.Request.Options?,
-                         listener: AuthFetchUserAttributeOperation.EventListener?) -> AuthFetchUserAttributeOperation
-
-    func update(userAttribute: AuthUserAttribute,
-                options: AuthUpdateUserAttributeOperation.Request.Options?,
-                listener: AuthUpdateUserAttributeOperation.EventListener?) -> AuthUpdateUserAttributeOperation
-
-    func update(userAttributes: [AuthUserAttribute],
-                options: AuthUpdateUserAttributesOperation.Request.Options?,
-                listener: AuthUpdateUserAttributesOperation.EventListener?) -> AuthUpdateUserAttributesOperation
-
-    func resendConfirmationCode(for attributeKey: AuthUserAttributeKey,
-                                options: AuthAttributeResendConfirmationCodeOperation.Request.Options?,
-                                listener: AuthAttributeResendConfirmationCodeOperation.EventListener?) -> AuthAttributeResendConfirmationCodeOperation
-
-    func confirm(userAttribute: AuthUserAttributeKey,
-                 confirmationCode: String,
-                 options: AuthConfirmUserAttributeOperation.Request.Options?,
-                 listener: AuthConfirmUserAttributeOperation.EventListener?) -> AuthConfirmUserAttributeOperation
-
-    func update(oldPassword: String,
-                to newPassword: String,
-                options: AuthChangePasswordOperation.Request.Options?,
-                listener: AuthChangePasswordOperation.EventListener?) -> AuthChangePasswordOperation
 }
